+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## [Unreleased]
+### Added
+- `FadeIn` effect.
+
 ### Fixed
 - `GaussianBlur` effect throwing a `TypeError` when applied to a movie or layer (the problem persisted).